Result when you Swipe Left
Basic outline of what happens when you swipe left;
A link to a page titled my questions;
A link to a page that takes you to your friends;
A link to a page titled settings;
Renny proposed that when we swipe right you will automatically see your
previous questions and etc, which I also agree with, but this is just a
basic template/outline. Can someone test to see if this works on a
mobile device? I can't test it on my phone.